PROJEKTI FULL/requirements.txt
Flask>=3.0.0
openai>=1.12.0
python-dotenv>=1.0.0
fpdf2>=2.7.10
